pyproject.toml
[project]
name = "detail-forge"
version = "0.1.0"
description = "AI-powered e-commerce product detail page generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "streamlit>=1.41.0",
    "playwright>=1.49.0",
    "Pillow>=11.0.0",
    "rembg>=2.0.62",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "selectolax>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
